+++ crates/analyzer/src/context/scope/conditional_scope.rs
@@ -10,6 +10,6 @@ pub struct IfConditionalScope<'a> {
     pub if_body_context: BlockContext<'a>,
     pub post_if_context: BlockContext<'a>,
     pub conditionally_referenced_variable_ids: HashSet<String>,
-    pub assigned_in_conditional_variable_ids: HashMap<String, usize>,
+    pub assigned_in_conditional_variable_ids: HashMap<String, u32>,
     pub entry_clauses: Vec<Clause>,
 }
