+++ crates/formatter/src/internal/format/statement.rs
@@ -47,7 +47,7 @@ fn print_statement_slice<'a>(f: &mut FormatterState<'a>, stmts: &[&'a Statement]
             let line = f.file.line_number(offset);
 
             if let Some(line_start_offset) = f.file.get_line_start_offset(line) {
-                let c = &f.file.contents[line_start_offset..offset];
+                let c = &f.file.contents[line_start_offset as usize..offset as usize];
                 let ws = c.chars().take_while(|c| c.is_whitespace()).collect::<String>();
                 if !ws.is_empty() {
                     let mut j = i + 1;
