+++ crates/syntax/src/comments/mod.rs
@@ -14,13 +14,13 @@ pub mod docblock;
 /// line from the start of the entire trivia text (including `/**`, `//`, etc.),
 /// and the `&str` is the cleaned line content.
 #[inline]
-pub fn comment_lines<'a>(trivia: &Trivia, interner: &'a ThreadedInterner) -> Vec<(usize, &'a str)> {
+pub fn comment_lines<'a>(trivia: &Trivia, interner: &'a ThreadedInterner) -> Vec<(u32, &'a str)> {
     let full_text = interner.lookup(&trivia.value);
     let (content_start_offset, content_end_offset) = match trivia.kind {
-        TriviaKind::MultiLineComment => (2, full_text.len() - 2),
-        TriviaKind::DocBlockComment => (3, full_text.len() - 2),
-        TriviaKind::SingleLineComment => (2, full_text.len()),
-        TriviaKind::HashComment => (1, full_text.len()),
+        TriviaKind::MultiLineComment => (2u32, full_text.len() as u32 - 2),
+        TriviaKind::DocBlockComment => (3u32, full_text.len() as u32 - 2),
+        TriviaKind::SingleLineComment => (2u32, full_text.len() as u32),
+        TriviaKind::HashComment => (1u32, full_text.len() as u32),
         TriviaKind::WhiteSpace => return vec![],
     };
 
@@ -29,13 +29,13 @@ pub fn comment_lines<'a>(trivia: &Trivia, interner: &'a ThreadedInterner) -> Vec
         return vec![];
     }
 
-    let content = &full_text[content_start_offset..content_end_offset];
+    let content = &full_text[content_start_offset as usize..content_end_offset as usize];
 
     let mut lines = Vec::new();
 
     for line in content.lines() {
         // Calculate the offset of the line relative to the start of the `content` slice.
-        let relative_line_offset = (line.as_ptr() as usize) - (content.as_ptr() as usize);
+        let relative_line_offset = (line.as_ptr() as u32) - (content.as_ptr() as u32);
         // Add the initial offset to get the position from the start of the entire trivia string.
         let offset_in_trivia = content_start_offset + relative_line_offset;
 
@@ -46,7 +46,7 @@ pub fn comment_lines<'a>(trivia: &Trivia, interner: &'a ThreadedInterner) -> Vec
         };
 
         // Calculate how many bytes were trimmed from the start of the original line slice.
-        let trimmed_bytes = (cleaned_line.as_ptr() as usize) - (line.as_ptr() as usize);
+        let trimmed_bytes = (cleaned_line.as_ptr() as u32) - (line.as_ptr() as u32);
         let final_offset = offset_in_trivia + trimmed_bytes;
 
         lines.push((final_offset, cleaned_line));
